use None option, precompress in non-sequential case
